aws-enclave/requirements.txt
llama-cpp-python==0.3.4
pydantic>=2.10,<3.0
cryptography>=44.0,<45.0
cbor2>=5.6,<6.0
scipy>=1.14,<2.0
